Moved ParameterizedUnitary into the utils folder
