+++ lightworks/sdk/circuit/__init__.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .parameterized_unitary import ParameterizedUnitary
 from .parameters import Parameter, ParameterDict
 from .photonic_circuit import PhotonicCircuit
 from .unitary import Unitary
