+++ lightworks/sdk/circuit/photonic_circuit.py
@@ -29,9 +29,11 @@ from lightworks.sdk.utils.exceptions import (
     CircuitCompilationError,
     ModeRangeError,
 )
+from lightworks.sdk.utils.parameterized_unitary import ParameterizedUnitary
 from lightworks.sdk.visualisation import Display
 
-from .circuit_utils import (
+from .parameters import Parameter
+from .photonic_circuit_utils import (
     add_empty_mode_to_circuit_spec,
     add_modes_to_circuit_spec,
     check_loss,
@@ -39,8 +41,6 @@ from .circuit_utils import (
     convert_non_adj_beamsplitters,
     unpack_circuit_spec,
 )
-from .parameterized_unitary import ParameterizedUnitary
-from .parameters import Parameter
 from .photonic_compiler import CompiledPhotonicCircuit
 from .photonic_components import (
     Barrier,
