+++ lightworks/sdk/circuit/photonic_circuit_utils.py
@@ -27,7 +27,8 @@ import sympy as sp
 from multimethod import multimethod
 from numpy.typing import NDArray
 
-from .parameterized_unitary import ParameterizedUnitary
+from lightworks.sdk.utils.parameterized_unitary import ParameterizedUnitary
+
 from .parameters import Parameter
 from .photonic_components import (
     Barrier,
